requirements.txt
attrs==19.3.0
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
docker==4.2.2
Flask==1.1.2
Flask-Cors==3.0.8
idna==2.10
importlib-metadata==1.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.3.0
mysql-connector-python==8.0.20
packaging==20.4
pluggy==0.13.1
protobuf==3.12.2
py==1.8.1
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.3
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.10
wcwidth==0.2.4
websocket-client==0.57.0
Werkzeug==1.0.1
zipp==3.1.0
